requirements.txt
fastapi~=0.115.0
requests~=2.32.3
python-dotenv~=1.0.1